DPR/requirements.txt
absl-py==2.3.1
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
faiss-cpu==1.13.1
filelock==3.20.0
fsspec==2025.12.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==11.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1+cu128
torchvision==0.24.1+cu128
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
